refactor: pageページからロジックを分離
